scripts/requirements.txt
streamlit==1.28.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.1
chromadb==0.4.15
TTS==0.20.6
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
sounddevice==0.4.6
soundfile==0.12.1
pyaudio==0.2.11
pydub==0.25.1
matplotlib==3.7.2
plotly==5.17.0
python-dotenv==1.0.0
streamlit-webrtc==0.47.1
av==10.0.0
